reorganize legacy
+ Removed SRML.Compatibility, and moved all of the legacy content to SRML.API

I don't love this move, but due to how dependencies work, the existence of SRML.Compatibility causes a circular dependency. I didn't pick it up because of how I was previously developing the project, but it's now a rather large issue.

I know this broke literally everything, but there's not much of a fix for that other than finishing the API rewrite.